64-bit support for db object containers in the Java wrapper

+++ java/jni/src/whitedb/holder/Database.java
@@ -29,5 +29,5 @@
 package whitedb.holder;
 
 public class Database {
-    public int pointer;
+    public long pointer;
 }
